jsAPI mostly complete; looking for tree table issue

+++ extensions/barrage/src/main/java/io/deephaven/extensions/barrage/BarrageSnapshotOptions.java
@@ -54,7 +54,7 @@ public abstract class BarrageSnapshotOptions implements StreamReaderOptions {
 
     @Override
     @Default
-    public int previewListLengthLimit() {
+    public long previewListLengthLimit() {
         return 0;
     }
 
@@ -111,7 +111,7 @@ public abstract class BarrageSnapshotOptions implements StreamReaderOptions {
          * @param previewListLengthLimit the magnitude of the number of elements to include in a preview list
          * @return this builder
          */
-        Builder previewListLengthLimit(int previewListLengthLimit);
+        Builder previewListLengthLimit(long previewListLengthLimit);
 
         /**
          * @return a new BarrageSnapshotOptions instance
